+++ src/sn/GrpLayer.ts
@@ -7,97 +7,50 @@
 
 import {Layer} from './Layer';
 
-import {CmnLib, int, IEvtMng, argChk_Num, getFn, getExt, argChk_Boolean} from './CmnLib';
+import {CmnLib, argChk_Num} from './CmnLib';
 import {HArg} from './Grammar';
-import {IMain, IVariable, SYS_DEC_RET} from './CmnInterface';
+import {IMain, IVariable} from './CmnInterface';
 import {Config} from './Config';
 import {SysBase} from './SysBase';
 import {EventListenerCtn} from './EventListenerCtn';
 import {SoundMng} from './SoundMng';
 import {IMakeDesignCast} from './LayerMng';
 import {GrpLayDesignCast} from './DesignCast';
-import {DebugMng} from './DebugMng';
-import {SEARCH_PATH_ARG_EXT} from './ConfigBase';
+import {SpritesMng} from './SpritesMng';
 
-import {Sprite, Container, Texture, BLEND_MODES, utils, Loader, LoaderResource, AnimatedSprite, Rectangle, RenderTexture, Application} from 'pixi.js';
-
-export type IFncCompSpr = (sp: Sprite)=> void;
-
-interface Iface {
-	fn			: string;
-	dx			: number;
-	dy			: number;
-	blendmode	: number;
-};
-interface Ihface { [name: string]: Iface; };
-
-interface IResAniSpr {
-	aTex	: Texture[];
-	meta	: {
-		animationSpeed? :number;
-	};
-}
+import {Sprite, AnimatedSprite, RenderTexture, Application} from 'pixi.js';
 
 
 export class GrpLayer extends Layer {
 	static	readonly	#elc	= new EventListenerCtn;
 
-	static	#hFace	: Ihface	= {};
-
 	static	#main	: IMain;
-	static	#cfg	: Config;
 	static	#appPixi: Application;
-	static	#sys	: SysBase;
-	static	#val	: IVariable;
-	static	#glbVol	= 1;
-	static	#movVol	= 1;
 	static	init(main: IMain, cfg: Config, appPixi: Application, sys: SysBase, sndMng: SoundMng, val: IVariable): void {
 		GrpLayer.#main = main;
-		GrpLayer.#cfg = cfg;
 		GrpLayer.#appPixi = appPixi;
-		GrpLayer.#sys = sys;
-		GrpLayer.#val = val;
-		const fnc = ()=> {
-			const vol = GrpLayer.#glbVol * GrpLayer.#movVol;
-			for (const v of Object.values(GrpLayer.hFn2VElm)) v.volume = vol;
-		};
-		sndMng.setNoticeChgVolume(
-			vol=> {GrpLayer.#glbVol = vol; fnc()},
-			vol=> {GrpLayer.#movVol = vol; fnc()}
-		);
-
-		if (GrpLayer.#sys.crypto) GrpLayer.#dec2cache = GrpLayer.#dec2cache4Cripto;
+		SpritesMng.init(cfg, val, sys, main, sndMng);
 	}
-	static	#evtMng	: IEvtMng;
-	static	setEvtMng(evtMng: IEvtMng) {GrpLayer.#evtMng = evtMng}
-	static	destroy() {
-		GrpLayer.#elc.clear();
-
-		GrpLayer.#hFace	= {};
-		GrpLayer.hFn2ResAniSpr	= {};
-//		GrpLayer.ldrHFn	= {};
-		GrpLayer.hFn2VElm	= {};
-	}
+	static	destroy() {GrpLayer.#elc.clear(); SpritesMng.destroy()}
 
 	readonly	#idc	= new GrpLayDesignCast(this.spLay, this);
 	constructor() {
 		super();
 		if (CmnLib.isDbg) {
-			this.setSp = sp=> this.#idc.setSp(sp);
+			this.#setIdcSp = sp=> this.#idc.setSp(sp);
 			this.cvsResize = ()=> {super.cvsResize(); this.#idc.cvsResize()}
 		}
 	}
-	private	setSp(_sp: Sprite) {}
+	#setIdcSp	: (sp: Sprite)=> void	= ()=> {};
 
 	#csvFn		= '';
 	#sBkFn		= '';
 	#sBkFace	= '';
-	static	hFn2ResAniSpr	: {[name: string]:	IResAniSpr} = {};
-	override readonly	lay = (hArg: HArg)=> this.laySub(hArg, isStop=> {
+	override readonly	lay = (hArg: HArg)=> this.#laySub(hArg, isStop=> {
 		if (isStop) GrpLayer.#main.resume();
 	});
-	private	laySub(hArg: HArg, resolve: (isStop: boolean)=> void): boolean {
-		GrpLayer.#stopVideo(this.#sBkFn);
+	#laySub(hArg: HArg, resolve: (isStop: boolean)=> void): boolean {
+		SpritesMng.stopVideo(this.#sBkFn);
 
 		const {fn, face = ''} = hArg;
 		this.#idc.sethArg(hArg);
@@ -122,7 +75,8 @@ export class GrpLayer extends Layer {
 		hArg.dx = 0;
 		hArg.dy = 0;
 
-		return GrpLayer.csv2Sprites(
+		this.#sps.destroy();
+		this.#sps = new SpritesMng(
 			this.#csvFn = fn + (face ? ','+ face : ''),
 			this.spLay,
 			sp=> {
@@ -135,245 +89,20 @@ export class GrpLayer extends Layer {
 				Layer.setXY(sp, hArg, this.spLay, true);
 			//	if (hArg.page === 'fore') this.rsvEvent(sp);	// ======
 					// [lay page=fore]のみswfアニメ終了イベント発生
-				this.setSp(sp);
+				this.#setIdcSp(sp);
 			},
 			isStop=> {
 				Layer.setBlendmodeParentOnly(this.spLay, hArg);
 				resolve(isStop);
 			}
 		);
+		return this.#sps.ret;
 	}
+	#sps	= new SpritesMng;
 	#width	= 0;
 	#height	= 0;
 	override	get	width() {return this.#width}
 	override	get	height() {return this.#height}
-
-//	static #ldrHFn: {[name: string]: 1} = {};
-	static csv2Sprites(csv: string, parent: Container | undefined, fncFirstComp: IFncCompSpr, fncAllComp: (isStop: boolean)=> void = ()=> {}): boolean {
-		// Data URI
-		let needLoad = false;
-		if (csv.slice(0, 5) === 'data:') {
-			const fnc = ()=> {
-				const sp = Sprite.from(csv);
-				parent?.addChild(sp);
-				fncFirstComp(sp);
-				fncAllComp(needLoad);
-			};
-			if (csv in utils.TextureCache) fnc();
-			else {needLoad = true; (new Loader).add(csv, csv).load(fnc)}
-
-			return needLoad;
-		}
-
-		const aComp: {fn: string, fnc: IFncCompSpr}[] = [];
-		const ldr = new Loader;
-		csv.split(',').forEach((fn0, i)=> {
-			if (! fn0) throw 'face属性に空要素が含まれます';
-
-			// 差分絵を重ねる
-			const {dx, dy, blendmode, fn} = GrpLayer.#hFace[fn0] || {
-				fn	: fn0,
-				dx	: 0,
-				dy	: 0,
-				blendmode	: BLEND_MODES.NORMAL
-			};
-			const fnc = (i === 0) ?fncFirstComp :(sp: Sprite)=> {
-				sp.x = dx;
-				sp.y = dy;
-				sp.blendMode = blendmode;
-			};
-			aComp.push({fn, fnc});
-
-			if (fn in GrpLayer.hFn2ResAniSpr) return;
-			if (fn in utils.TextureCache) return;
-			if (fn in Loader.shared.resources) return;
-//			if (fn in GrpLayer.ldrHFn) {
-				// ここに来るという中途半端な状態がある。お陰で警告が出てしまう
-// 以下の試みは効かない
-//	Texture.removeFromCache(fn);
-//	delete utils.TextureCache[fn];
-//	delete Loader.shared.resources[fn];
-				// return;	// これは厳禁、御法度。
-					// 画像ボタンや文字ボタン背景で同じ画像を、間を置かずロードした場合に最初一つしか表示されなくなる。以下は確認用
-					// http://localhost:8082/index.html?cur=ch_button
-//			}
-//			GrpLayer.ldrHFn[fn] = 1;
-
-			needLoad = true;
-			const url = GrpLayer.#cfg.searchPath(fn, SEARCH_PATH_ARG_EXT.SP_GSM);
-			const xt = this.#sys.crypto
-			? {xhrType: (url.slice(-5) === '.json')
-				? LoaderResource.XHR_RESPONSE_TYPE.TEXT
-				: LoaderResource.XHR_RESPONSE_TYPE.BUFFER}
-			: {};
-			ldr.add({...xt, name: fn, url});
-		});
-
-		const fncLoaded = (hRes: {[fn: string]: LoaderResource})=> {
-			for (const {fn, fnc} of aComp) {
-				const sp = GrpLayer.#mkSprite(fn, hRes);
-				/**/sp.name = fn;
-				parent?.addChild(sp);
-				fnc(sp);
-			}
-			fncAllComp(needLoad);
-		}
-		if (needLoad) {
-			ldr.use((res, next)=> {
-				this.#sys.dec(res.extension, res.data)
-				.then(r=> GrpLayer.#dec2cache(r, res, ()=> next?.()))
-				.catch(e=> this.#main.errScript(`Graphic ロード失敗です fn:${res.name} ${e}`, false));
-			})
-			.load((_ldr, hRes)=> fncLoaded(hRes));
-		}
-		else fncLoaded({});
-
-		return needLoad;
-	}
-	static #dec2cache = (_r: SYS_DEC_RET, res: any, next: ()=> void)=> {
-		switch (res.type) {
-			case LoaderResource.TYPE.JSON:
-				// アニメ登録
-				const aFn: string[] = res.spritesheet._frameKeys;
-				GrpLayer.#sortAFrameName(aFn);
-				GrpLayer.hFn2ResAniSpr[res.name] = {
-					aTex: aFn.map(fn=> Texture.from(fn)),
-					meta: res.data.meta,
-				};
-				break;
-
-			case LoaderResource.TYPE.VIDEO:
-				const hve = res.data as HTMLVideoElement;
-				hve.volume = GrpLayer.#glbVol;
-				GrpLayer.hFn2VElm[res.name] = GrpLayer.#charmVideoElm(hve);
-		}
-		next();
-	}
-	static #sortAFrameName(aFn: string[]): string[] {
-		const a_base_name = /([^\d]+)\d+\.(\w+)/.exec(aFn[0]);
-		if (! a_base_name) return [];
-
-		const is = a_base_name[1].length;
-		const ie = -a_base_name[2].length -1;
-		return aFn.sort((a, b)=>
-			int(a.slice(is, ie)) > int(b.slice(is, ie)) ?1 :-1
-		);
-	}
-	static #dec2cache4Cripto(r: SYS_DEC_RET, res: any, next: ()=> void) {
-		res.data = r;
-		if (res.extension === 'bin') {
-			if (r instanceof HTMLImageElement) {
-				res.texture = Texture.fromLoader(r, res.url, res.name);
-				//Texture.addToCache(Texture.from(r), res.name);
-				// res.texture = Texture.from(r);
-					// でも良いが、キャッシュ追加と、それでcsv2Sprites()内で使用するので
-				res.type = LoaderResource.TYPE.IMAGE;
-//				URL.revokeObjectURL(r.src);	// TODO: キャッシュ破棄
-			}
-			else if (r instanceof HTMLVideoElement) {
-				r.volume = GrpLayer.#glbVol;
-				GrpLayer.hFn2VElm[res.name] = GrpLayer.#charmVideoElm(r);
-
-				res.type = LoaderResource.TYPE.VIDEO;
-//				URL.revokeObjectURL(r.src);
-			}
-		}
-		if (res.extension !== 'json') {next(); return}
-
-		// アニメ登録
-		if (typeof r !== 'string') {next(); return}
-
-		const {meta, frames} = res.data = JSON.parse(r);
-		res.type = LoaderResource.TYPE.JSON;
-		if (! meta?.image) {next(); return}
-
-		const fn = getFn(meta.image);
-		const url = GrpLayer.#cfg.searchPath(fn, SEARCH_PATH_ARG_EXT.SP_GSM);
-		(new Loader)
-		.use((res2, next2)=> {
-			this.#sys.dec(res2.extension, res2.data)
-			.then(r2=> {
-				res2.data = r2;
-				if (r2 instanceof HTMLImageElement) {
-					res2.type = LoaderResource.TYPE.IMAGE;
-					URL.revokeObjectURL(r2.src);
-				}
-				next2?.();
-			})
-			.catch(e=> this.#main.errScript(`Graphic ロード失敗です dec2res4Cripto fn:${res2.name} ${e}`, false));
-		})
-		.add({name: fn, url, xhrType: LoaderResource.XHR_RESPONSE_TYPE.BUFFER})
-		.load((ldr, _hRes)=> {
-			for (const {data} of Object.values(ldr.resources)) {
-				const {baseTexture} = Texture.from(data);
-				const aFr = Object.values(<{[nm: string]: {
-					frame: {x: number; y: number; w: number; h: number;}
-				}}>frames);
-				GrpLayer.hFn2ResAniSpr[res.name] = {
-					aTex: aFr.map(({frame: {x, y, w, h}})=> new Texture(
-						baseTexture,
-						new Rectangle(x, y, w, h),
-					)),
-					meta,
-				};
-			}
-
-			next();
-		});
-	}
-	static #charmVideoElm(v: HTMLVideoElement): HTMLVideoElement {
-		// 【PixiJS】iOSとChromeでAutoPlay可能なビデオSpriteの設定 - Qiita https://qiita.com/masato_makino/items/8316e7743acac514e361
-		// v.muted = true;// Chrome対応：自動再生を許可。ないと再開時に DOMException
-		if (GrpLayer.#val.getVal('const.sn.needClick2Play')) {
-			// ブラウザ実行で、クリックされるまで音声再生が差し止められている状態か。なにかクリックされれば falseになる
-			DebugMng.trace_beforeNew(`[lay系] ${DebugMng.strPos()}未クリック状態で動画を自動再生します。音声はミュートされます`, 'W');
-			v.muted = true;
-		}
-		v.setAttribute('playsinline', '');	// iOS対応
-		return v;
-	}
-	static #mkSprite(fn: string, hRes: {[fn: string]: LoaderResource}): Sprite {
-		const ras = GrpLayer.hFn2ResAniSpr[fn];
-		if (ras) {
-			const asp = new AnimatedSprite(ras.aTex);
-			asp.animationSpeed = ras.meta.animationSpeed ?? 1.0;
-			asp.play();
-			return asp;
-		}
-		if (fn in utils.TextureCache) return Sprite.from(fn);
-		if (fn in GrpLayer.hFn2VElm) return Sprite.from(GrpLayer.hFn2VElm[fn]);
-
-		return (fn in hRes) ?new Sprite(hRes[fn].texture) :new Sprite;
-	}
-	static hFn2VElm	: {[name: string]: HTMLVideoElement} = {};
-	static wv(hArg: HArg) {
-		// 動画ファイル名指定でいいかなと。なぜなら「ループで再生しつつ」
-		// 同ファイル名の別の動画の再生は待ちたい、なんて状況は普通は無いだろうと
-		const {fn} = hArg;
-		if (! fn) throw 'fnは必須です';
-		const hve = GrpLayer.hFn2VElm[fn];
-		if (! hve || hve.loop) return false;
-
-		if (this.#evtMng.isSkipping() || hve.ended) {GrpLayer.#stopVideo(fn); return false}
-
-		const fncBreak = ()=> this.#evtMng.breakLimitedEvent();	// waitEvent 使用者の通常 break 時義務
-		hve.addEventListener('ended', fncBreak, {once: true, passive: true});
-
-		const stop = argChk_Boolean(hArg, 'stop', true);
-		return GrpLayer.#evtMng.waitEvent(hArg, ()=> {
-			hve.removeEventListener('ended', fncBreak);
-			if (stop) GrpLayer.#stopVideo(fn);
-			fncBreak()
-		});
-	}
-	static	#stopVideo(fn: string) {
-		const hve = GrpLayer.hFn2VElm[fn];
-		if (! hve) return;
-
-		delete GrpLayer.hFn2VElm[fn];
-		hve.pause();
-		hve.currentTime = hve.duration;
-	}
 
 
 	override	renderStart() {
@@ -416,49 +145,6 @@ export class GrpLayer extends Layer {
 	}
 
 
-	static	#REG_REP_PRM = /\?([^?]+)$/;	// https://regex101.com/r/ZUnoFq/1
-	static	loadPic2Img(src: string, img: HTMLImageElement, onload?: (img2: HTMLImageElement)=> void) {
-		const srcNoPrm = src.replace(GrpLayer.#REG_REP_PRM, '');
-		const prmSrc = (src === srcNoPrm) ?'' :src.slice(srcNoPrm.length);
-		const oUrl = this.#hEncImgOUrl[src];
-		if (oUrl) {img.src = oUrl; return}
-
-		const aImg = this.#hAEncImg[src];
-		if (aImg) {aImg.push(img); return}
-		this.#hAEncImg[src] = [img];
-
-		const url2 = GrpLayer.#cfg.searchPath(srcNoPrm, SEARCH_PATH_ARG_EXT.SP_GSM);
-		const ld2 = (new Loader)
-		.add({name: src, url: url2, xhrType: LoaderResource.XHR_RESPONSE_TYPE.BUFFER,});
-		if (GrpLayer.#sys.crypto && getExt(url2) === 'bin') ld2.use((res, next)=> {
-			GrpLayer.#sys.dec(res.extension, res.data)
-			.then(r=> {
-				if (res.extension !== 'bin') {next?.(); return}
-				res.data = r;
-				if (r instanceof HTMLImageElement) {
-					res.type = LoaderResource.TYPE.IMAGE;
-				}
-				next?.();
-			})
-			.catch(e=> GrpLayer.#main.errScript(`GrpLayer loadPic ロード失敗です fn:${res.name} ${e}`, false));
-		})
-		ld2.load((_ldr, hRes)=> {
-			for (const [s2, {data: {src}}] of Object.entries(hRes)) {
-				const u2 = this.#hEncImgOUrl[s2]
-				= src + (src.slice(0, 5) === 'blob:' ?'' :prmSrc);
-				for (const i of this.#hAEncImg[s2]) {
-					i.src = u2;
-					if (onload) i.onload = ()=> onload(i);
-				}
-				delete this.#hAEncImg[s2];
-			//	URL.revokeObjectURL(u2);// 画面遷移で毎回再生成するので
-			}
-		});
-	}
-	static	#hAEncImg		: {[name: string]: HTMLImageElement[]}	= {};
-	static	#hEncImgOUrl	: {[name: string]: string}				= {};
-
-
 	setPos(hArg: HArg): void {
 		Layer.setXY(
 			(this.spLay.children.length === 0) ?this.spLay :this.spLay.children[0],
@@ -468,39 +154,22 @@ export class GrpLayer extends Layer {
 		);
 	}
 
-	static	add_face(hArg: HArg): boolean {
-		const {name} = hArg;
-		if (! name) throw 'nameは必須です';
-		if (name in GrpLayer.#hFace) throw '一つのname（'+ name +'）に対して同じ画像を複数割り当てられません';
-
-		const {fn = name} = hArg;
-		GrpLayer.#hFace[name] = {
-			fn,
-			dx: argChk_Num(hArg, 'dx', 0),
-			dy: argChk_Num(hArg, 'dy', 0),
-			blendmode: Layer.getBlendmodeNum(hArg.blendmode || '')
-		};
-
-		return false;
-	}
-	static	clearFace2Name(): void {GrpLayer.#hFace = {}}
-
 	// アニメ・動画を含むか
 	override get containMovement(): boolean {
 		if (this.#csvFn === '') return false;
 
 		const c = this.spLay.children;
 		return this.#csvFn.split(',').some(
-			(fn, i)=> c[i] instanceof AnimatedSprite || GrpLayer.hFn2VElm[fn]
+			(fn, i)=> c[i] instanceof AnimatedSprite || SpritesMng.getHFn2VElm(fn)
 		);
 	}
 
 	override clearLay(hArg: HArg): void {
 		super.clearLay(hArg);
-		for (const c of this.spLay.removeChildren()) c.destroy();
-		this.#sBkFn	= '';
-		this.#sBkFace= '';
-		this.#csvFn	= '';
+		this.#sps.destroy();
+		this.#sBkFn		= '';
+		this.#sBkFace	= '';
+		this.#csvFn		= '';
 	}
 	override readonly record = ()=> ({...super.record(),
 		sBkFn		: this.#sBkFn,
@@ -516,7 +185,7 @@ export class GrpLayer extends Layer {
 			return;
 		}
 
-		aPrm.push(new Promise(re=> this.laySub(
+		aPrm.push(new Promise(re=> this.#laySub(
 			{fn: hLay.sBkFn, face: hLay.sBkFace, left: hLay.x, top: hLay.y, alpha: hLay.alpha, blendmode: Layer.getNum2Blendmode(hLay.blendMode), rotation: hLay.rotation, scale_x: hLay.scale_x, scale_y: hLay.scale_y},
 			_isStop=> {this.spLay.position.set(hLay.x, hLay.y); re()},
 				// Layer.setXY()の後に再度移動